move to core/dtypes/cast.py
